// src/app/features/health-records/models/update-health-record.dto.ts
export type UpdateHealthRecordDto = {
  petName?: string | null;
  breed?: string | null;
  sex?: string | null;
  birthDate?: string | null;
  currentWeight?: number | null;
  color?: string | null;
  identificationNumber?: string | null;
  tattooNumber?: string | null;
  allergy?: string | null;
};